Mudando nome do caminho (listarCpf -> listar) e mudando tipo do cpf

+++ Models/Cliente.cs
@@ -1,21 +1,23 @@
 
-namespace EcommerceAPI.Models{
-    
-    public class Cliente{
+namespace EcommerceAPI.Models
+{
+
+    public class Cliente
+    {
         public int Id { get; set; }
 
         public string NomeTutor { get; set; }
 
-        public int CPF { get; set; }
+        public long CPF { get; set; }
 
         public string Telefone { get; set; }
-        
+
         public string NomePet { get; set; }
 
         public string Especie { get; set; }
 
         public int Idade { get; set; }
-        
+
         public string Porte { get; set; }
     }
 }
